backend/requirements.txt
# Confluence AI Chatbot — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1

# ── AI Providers (install only the ones you use) ──────────
groq==0.11.0                    # Groq (free Llama)
openai==1.51.0                  # OpenAI GPT-4o
anthropic==0.34.0               # Anthropic Claude
google-generativeai==0.8.0      # Google Gemini
httpx==0.27.0                   # Ollama (uses httpx directly)

# ── Embeddings (free, runs locally) ──────────────────────
sentence-transformers==3.2.1
torch==2.4.1                    # Required by sentence-transformers

# ── Vector Database ───────────────────────────────────────
qdrant-client==1.11.3

# ── Document Parsers ──────────────────────────────────────
PyMuPDF==1.24.11                # PDF parsing (fitz)
python-docx==1.1.2              # Word .docx parsing
openpyxl==3.1.5                 # Excel .xlsx parsing
beautifulsoup4==4.12.3          # HTML → text (Confluence pages)
lxml==5.3.0                     # HTML parser for BeautifulSoup

# ── Utilities ─────────────────────────────────────────────
tqdm==4.67.0                    # Progress bars during embedding
